Created the presentation layer, move main method and add menu class.
