+++ service/StudentService.java
@@ -33,16 +33,5 @@ public class StudentService {
         studentDao.delete(studentId);
     }
 
-    public void showMenu(){
-        System.out.println("Available Menu Items: ");
-        System.out.println("1: Add Student");
-        System.out.println("2: Update Student");
-        System.out.println("3: Get Students");
-        System.out.println("""
-                4: Delete Student
 
-                """);
-        System.out.println("Which function do you want to execute>");
-        System.out.print("Menu Item: ");
-    }
 }
